address feedback from @cjdsellers
