+++ nautilus_trader/indicators/bid_ask_min_max.pxd
@@ -10,11 +10,12 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+from cpython.datetime cimport datetime
 from cpython.datetime cimport timedelta
 
 from nautilus_trader.indicators.base.indicator cimport Indicator
-from nautilus_trader.indicators.utils.windowed_min_max_prices cimport WindowedMinMaxPrices
 from nautilus_trader.model.identifiers cimport Symbol
+from nautilus_trader.model.objects cimport Price
 from nautilus_trader.model.tick cimport QuoteTick
 
 
@@ -32,3 +33,25 @@ cdef class BidAskMinMax(Indicator):
 
     cpdef void handle_quote_tick(self, QuoteTick tick) except *
     cpdef void reset(self)
+
+
+cdef class WindowedMinMaxPrices:
+    """
+    Over the course of a defined lookback window, efficiently keep track
+    of the min/max values currently in the window.
+    """
+
+    cdef readonly timedelta lookback
+
+    cdef readonly Price min_price
+    cdef readonly Price max_price
+
+    cdef object _min_prices
+    cdef object _max_prices
+
+    cpdef void add_price(self, datetime ts, Price price)
+    cpdef void reset(self)
+
+    cdef inline void _expire_stale_prices_by_cutoff(self, object ts_prices, datetime cutoff)
+    cdef inline void _add_min_price(self, datetime ts, Price price)
+    cdef inline void _add_max_price(self, datetime ts, Price price)
