+++ tests/unit_tests/indicators/test_max_bid_min_ask.py
@@ -17,6 +17,7 @@ import unittest
 import pytz
 
 from nautilus_trader.indicators.bid_ask_min_max import BidAskMinMax
+from nautilus_trader.indicators.bid_ask_min_max import WindowedMinMaxPrices
 from nautilus_trader.model.identifiers import Symbol
 from nautilus_trader.model.identifiers import Venue
 from nautilus_trader.model.objects import Price
@@ -39,11 +40,11 @@ class BidAskMinMaxTests(unittest.TestCase):
         self.assertEqual(None, indicator.asks.max_price)
         self.assertEqual(False, indicator.initialized)
 
-    def test_can_expire_items(self):
+    def test_handle_quote_tick(self):
         # Arrange
         indicator = BidAskMinMax(self.symbol, timedelta(minutes=5))
 
-        # Act + Assert
+        # Act
         indicator.handle_quote_tick(
             QuoteTick(
                 self.symbol,
@@ -54,11 +55,6 @@ class BidAskMinMaxTests(unittest.TestCase):
                 datetime(2020, 1, 1, 0, 0, 0, tzinfo=pytz.utc),
             )
         )
-        self.assertEqual(Price(1.0, 0), indicator.bids.min_price)
-        self.assertEqual(Price(1.0, 0), indicator.bids.max_price)
-        self.assertEqual(Price(2.0, 0), indicator.asks.min_price)
-        self.assertEqual(Price(2.0, 0), indicator.asks.max_price)
-
         # 5 min later (still in the window)
         indicator.handle_quote_tick(
             QuoteTick(
@@ -70,24 +66,92 @@ class BidAskMinMaxTests(unittest.TestCase):
                 datetime(2020, 1, 1, 0, 5, 0, tzinfo=pytz.utc),
             )
         )
+
+        # Assert
         self.assertEqual(Price(0.9, 0), indicator.bids.min_price)
         self.assertEqual(Price(1.0, 0), indicator.bids.max_price)
         self.assertEqual(Price(2.0, 0), indicator.asks.min_price)
         self.assertEqual(Price(2.1, 0), indicator.asks.max_price)
 
+
+class WindowedMinMaxPricesTests(unittest.TestCase):
+    def test_can_instantiate(self):
+        # Arrange
+        instance = WindowedMinMaxPrices(timedelta(minutes=5))
+
+        # Act
+        # Assert
+        self.assertEqual(None, instance.min_price)
+        self.assertEqual(None, instance.max_price)
+
+    def test_can_add_price(self):
+        # Arrange
+        instance = WindowedMinMaxPrices(timedelta(minutes=5))
+
+        # Act
+        instance.add_price(
+            datetime(2020, 1, 1, 0, 0, 0, tzinfo=pytz.utc),
+            Price(1.0, 0),
+        )
+        # Assert
+        self.assertEqual(Price(1.0, 0), instance.min_price)
+        self.assertEqual(Price(1.0, 0), instance.max_price)
+
+    def test_can_add_multiple_prices(self):
+        # Arrange
+        instance = WindowedMinMaxPrices(timedelta(minutes=5))
+
+        # Act
+        instance.add_price(
+            datetime(2020, 1, 1, 0, 0, 0, tzinfo=pytz.utc),
+            Price(1.0, 0),
+        )
+        # 5 min later (still in the window)
+        instance.add_price(
+            datetime(2020, 1, 1, 0, 5, 0, tzinfo=pytz.utc),
+            Price(0.9, 0),
+        )
+
+        # Assert
+        self.assertEqual(Price(0.9, 0), instance.min_price)
+        self.assertEqual(Price(1.0, 0), instance.max_price)
+
+    def test_can_expire_items(self):
+        # Arrange
+        instance = WindowedMinMaxPrices(timedelta(minutes=5))
+
+        # Act
+        instance.add_price(
+            datetime(2020, 1, 1, 0, 0, 0, tzinfo=pytz.utc),
+            Price(1.0, 0),
+        )
+        # 5 min later (still in the window)
+        instance.add_price(
+            datetime(2020, 1, 1, 0, 5, 0, tzinfo=pytz.utc),
+            Price(0.9, 0),
+        )
         # Allow the first item to expire out
         # This also tests that the new tick is the new min/max
-        indicator.handle_quote_tick(
-            QuoteTick(
-                self.symbol,
-                Price(0.95, 0),
-                Price(2.05, 0),
-                Quantity(1),
-                Quantity(1),
-                datetime(2020, 1, 1, 0, 5, 1, tzinfo=pytz.utc),
-            )
+        instance.add_price(
+            datetime(2020, 1, 1, 0, 5, 1, tzinfo=pytz.utc),
+            Price(0.95, 0),
         )
-        self.assertEqual(Price(0.90, 0), indicator.bids.min_price)
-        self.assertEqual(Price(0.95, 0), indicator.bids.max_price)
-        self.assertEqual(Price(2.05, 0), indicator.asks.min_price)
-        self.assertEqual(Price(2.10, 0), indicator.asks.max_price)
+
+        # Assert
+        self.assertEqual(Price(0.90, 0), instance.min_price)
+        self.assertEqual(Price(0.95, 0), instance.max_price)
+
+    def test_reset(self):
+        # Arrange
+        instance = WindowedMinMaxPrices(timedelta(minutes=5))
+
+        # Act
+        instance.add_price(
+            datetime(2020, 1, 1, 0, 0, 0, tzinfo=pytz.utc),
+            Price(1.0, 0),
+        )
+        instance.reset()
+
+        # Assert
+        self.assertEqual(None, instance.min_price)
+        self.assertEqual(None, instance.max_price)
